Use samblaster to speed up duplicate marking
One of the reasons samblaster is fast is because the output of bwa mem can be
piped through it, as in

    bwa mem ... | samblaster | samtools sort - > out.bam

However, we typically work on libraries that were sequenced on multiple lanes,
which therefore come as multiple FASTQ pairs. BWA works on a single FASTQ pair
only. Previously, we would create separate BAM files for each FASTQ pair, merge
them afterwards and then run MarkDuplicates on the merged file, which is slow
and requires that it be run as a separate step that cannot be parallelized.

The solution in this commit is to introduce a wrapper `bwa_multifastq` for
bwa mem. It accepts multiple FASTQ pairs, runs bwa mem on each pair in turn and
sends the result to standard output, which can then be piped into samblaster.
The wrapper implements a lightweight SAM merger: It needs to avoid printing the
SAM header more than once, and also needs to make sure that the header that it
does print contains the @RG headers for all the FASTQ pairs.

Not only the MarkDuplicates process can be removed, but also the MergeBams
process. Additionally, the logic for deciding whether or not to merge BAM
files could be removed since we get this for free by using the wrapper.

Previously, MarkDuplicates would create the BAM index file `.bai'. This is
now also done within the MapReads in a streaming way by using some `mkfifo` and
`tee` magic.

The only downside is that the MarkDuplicates report is now gone, but possibly
samblaster can be coerced into producing something similar.