Avoid the MergeBam step
This commit introduces a wrapper script `bwa_multifastq` for bwa mem. It
accepts multiple FASTQ pairs, runs bwa mem on each pair in turn and sends
the result as a single stream to standard output, making sure that the
output is correct SAM.

The wrapper can forward almost all input unchanged, except that it needs to
avoid printing the SAM header more than once, and also it needs to make sure
that the one header that it does print contains the @RG headers for all the
FASTQ pairs.

The wrapper also automatically deduces a read group id from the flowcell id
stored in the input FASTQ files.

Not only can the MergeBam process be removed, but also the logic for
deciding whether or not to merge BAM files since we get this "for free" by
using the wrapper. MapReads now maps all FASTQs for a single sample in one
go.

The changes in this commit were also introduced by PR #242, but that PR went
further by replacing MarkDuplicates with samblaster. This commit keeps
MarkDuplicates and is hopefully less controversial.